// src/generated/kubb/chat-completions/zod/chatCompletionChunkSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { chatCompletionChunkChoiceSchema } from "./chatCompletionChunkChoiceSchema.ts";
import { completionUsageSchema } from "./completionUsageSchema.ts";
import { z } from "zod";

export const chatCompletionChunkSchema = z.object({
  id: z.string(),
  object: z.enum(["chat.completion.chunk"]),
  created: z.number().int(),
  model: z.string(),
  system_fingerprint: z.optional(z.union([z.string(), z.null()])),
  service_tier: z.optional(z.union([z.string(), z.null()])),
  choices: z.array(z.lazy(() => chatCompletionChunkChoiceSchema)),
  usage: z.optional(z.union([z.lazy(() => completionUsageSchema), z.null()])),
});