// src/generated/kubb/chat-completions/zod/chatCompletionSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { chatCompletionChoiceSchema } from "./chatCompletionChoiceSchema.ts";
import { completionUsageSchema } from "./completionUsageSchema.ts";
import { z } from "zod";

export const chatCompletionSchema = z.object({
  id: z.string(),
  object: z.enum(["chat.completion"]),
  created: z.number().int(),
  model: z.string(),
  system_fingerprint: z.optional(z.union([z.string(), z.null()])),
  service_tier: z.optional(z.union([z.string(), z.null()])),
  choices: z.array(z.lazy(() => chatCompletionChoiceSchema)),
  usage: z.optional(z.lazy(() => completionUsageSchema)),
});